requirements.txt
aiosqlite==0.17.0
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
filelock==3.7.0
Flask==2.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
iso8601==1.0.2
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pypika-tortoise==0.1.5
python-dateutil==2.8.2
pytz==2022.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
tldextract==3.3.0
tortoise-orm==0.19.1
Twisted==22.4.0
typing-extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
Werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0
